Refactored some methods and finished unit tests for Services

+++ src/main/java/com/doublemc/services/ToDoItemServiceBean.java
@@ -1,14 +1,15 @@
 package com.doublemc.services;
 
+import com.doublemc.customexceptions.ToDoItemNotFoundException;
+import com.doublemc.customexceptions.UserAccessException;
 import com.doublemc.domain.ToDoItem;
 import com.doublemc.domain.User;
 import com.doublemc.repositories.ToDoItemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.ResponseStatus;
 
 import javax.transaction.Transactional;
+import java.security.Principal;
 
 /**
  * Created by michal on 29.01.17.
@@ -17,43 +18,63 @@ import javax.transaction.Transactional;
 @Transactional
 public class ToDoItemServiceBean {
     private final ToDoItemRepository toDoItemRepository;
+    private final UserServiceBean userService;
 
     @Autowired
-    ToDoItemServiceBean(ToDoItemRepository toDoItemRepository) {
+    ToDoItemServiceBean(ToDoItemRepository toDoItemRepository, UserServiceBean userService) {
         this.toDoItemRepository = toDoItemRepository;
+        this.userService = userService;
     }
 
-    public ToDoItem addToDo(ToDoItem toDoItem, User user) {
-        ToDoItem newToDo = new ToDoItem(user, toDoItem.getTitle(), toDoItem.getDueDate());
+    public ToDoItem addToDo(ToDoItem toDoItem, Principal principal) {
+        User currentUser = userService.findLoggedInUser(principal);
+        ToDoItem newToDo = new ToDoItem(currentUser, toDoItem.getTitle(), toDoItem.getDueDate());
         return toDoItemRepository.save(newToDo);
     }
 
-    public ToDoItem editToDo(ToDoItem newToDoItem, ToDoItem oldToDoItem) {
-        oldToDoItem.setTitle(newToDoItem.getTitle());
-        oldToDoItem.setDueDate(newToDoItem.getDueDate());
-        return oldToDoItem;
+    public ToDoItem editToDo(ToDoItem newToDoItem, Long itemId, Principal principal) {
+        if (isToDoAccessible(itemId, principal)) {
+            ToDoItem toDoFromDb = findToDoItemById(itemId);
+            toDoFromDb.setTitle(newToDoItem.getTitle());
+            toDoFromDb.setDueDate(newToDoItem.getDueDate());
+            return toDoFromDb;
+        }
+        return null;
     }
 
-    public void deleteToDo(Long id) {
-        toDoItemRepository.delete(id);
+    public void deleteToDo(Long id, Principal principal) {
+        if(isToDoAccessible(id, principal)) {
+            toDoItemRepository.delete(id);
+        }
     }
 
-    public void completeToDo(ToDoItem toDoItem) {
-        toDoItem.setCompleted(true);
+    public void completeToDo(Long id, Principal principal) {
+        if(isToDoAccessible(id, principal)) {
+            findToDoItemById(id).setCompleted(true);
+        }
     }
 
-    public boolean toDoExists(Long id) {
+    private boolean isToDoAccessible(Long id, Principal principal) {
+        User currentUser = userService.findLoggedInUser(principal);
+        if (!toDoExists(id)) {
+            throw new ToDoItemNotFoundException();
+        }
+        ToDoItem toDoFromDb = findToDoItemById(id);
+        if (!canUserAccessToDo(toDoFromDb, currentUser)) {
+            throw new UserAccessException();
+        }
+        return true;
+    }
+
+    private boolean toDoExists(Long id) {
         return toDoItemRepository.findOne(id) != null;
     }
 
-    public boolean canUserAccessToDo(ToDoItem toDoItem, User user) {
-        if (toDoItem.getUser() == user) {
-            return true;
-        }
-        return false;
+    private boolean canUserAccessToDo(ToDoItem toDoItem, User user) {
+        return toDoItem.getUser() == user;
     }
 
-    public ToDoItem findToDoItemById(Long id) {
+    private ToDoItem findToDoItemById(Long id) {
         return toDoItemRepository.findOne(id);
     }
